Move Area, read_all, read_area_node to lib
Also clean up unused import warnings
Missing a good heuristic of setting max sizes
for heapless Vecs.
