+++ src/lib/device_tree/Cargo.toml
@@ -10,6 +10,7 @@ model = { path = "../../drivers/model"}
 num-derive = { version = "0.2", default-features = false }
 num-traits = { version = "0.2", default-features = false }
 wrappers = { path = "../../drivers/wrappers"}
+heapless = "0.6.0"
 
 [profile.release]
 opt-level = 'z'  # Optimize for size.
