+++ tools/layoutflash/src/main.rs
@@ -1,6 +1,5 @@
 use clap::Clap;
-use device_tree::{infer_type, Entry, FdtIterator, FdtReader, Type, MAX_NAME_SIZE};
-use model::{Driver, Result};
+use device_tree::{Area, read_area_node, Entry, FdtReader};
 use std::io;
 use std::io::{Seek, SeekFrom, Write};
 use std::process::exit;
@@ -11,52 +10,14 @@ use std::{
 use wrappers::SliceReader;
 
 // TODO: Move this struct to lib so it can be used at runtime.
-#[derive(Default, Debug)]
-struct Area {
-    description: String,
-    compatible: String,
-    // If not specified, it will be automatically computed based on previous areas (if this is
-    // first area, we start with 0).
-    offset: Option<u32>,
-    size: u32,
-    file: Option<String>,
-}
+// struct Area
 
 // TODO: Move to some common library.
-fn read_all(d: &dyn Driver) -> Vec<u8> {
-    let mut v = Vec::new();
-    v.resize(MAX_NAME_SIZE, 0);
-    // Safe to unwrap because SliceReader does not return an error.
-    let size = d.pread(v.as_mut_slice(), 0).unwrap();
-    v.truncate(size);
-    v
-}
+// fn read_all(d: &dyn Driver) -> Vec<u8> 
 
-fn read_area_node<D: Driver>(iter: &mut FdtIterator<D>) -> Result<Area> {
-    let mut area = Area {
-        ..Default::default()
-    };
-    while let Some(item) = iter.next()? {
-        match item {
-            Entry::StartNode { name: _ } => {
-                iter.skip_node()?;
-            }
-            Entry::EndNode => return Ok(area),
-            Entry::Property { name, value } => {
-                let data = read_all(&value);
-                match (name, infer_type(data.as_slice())) {
-                    ("description", Type::String(x)) => area.description = String::from(x),
-                    ("compatible", Type::String(x)) => area.compatible = String::from(x),
-                    ("offset", Type::U32(x)) => area.offset = Some(x),
-                    ("size", Type::U32(x)) => area.size = x,
-                    ("file", Type::String(x)) => area.file = Some(String::from(x)),
-                    (_, _) => {}
-                }
-            }
-        }
-    }
-    Ok(area)
-}
+
+// fn read_area_node<D: Driver>(iter: &mut FdtIterator<D>) -> Result<Area> {
+
 
 // TODO: Move this function to lib so it can be used at runtime.
 fn read_fixed_fdt(path: &Path) -> io::Result<Vec<Area>> {
